+++ CHANGELOG.md
@@ -155,3 +155,4 @@
  * Don't read curve information into the X.509 tea leaves where none exists
    * `CryptoSignature`s in X.509 are now indefinite length
  * Rename `CryptoAlgorithm` to `X509SignatureAlgorithm` to better describe what it is
+   * Rename `toCryptoAlgorithm` to `toX509SignatureAlgorithm` accordingly